add login

+++ src/main/java/com/dh/clinica/service/OdontologoService.java
@@ -1,11 +1,16 @@
 package com.dh.clinica.service;
 
+import com.dh.clinica.exceptions.BadRequestException;
+import com.dh.clinica.exceptions.ResourceNotFoundException;
 import com.dh.clinica.model.Odontologo;
 import com.dh.clinica.model.OdontologoDTO;
 import com.dh.clinica.repository.impl.OdontologoRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 
 
 import java.util.HashSet;
@@ -17,6 +22,7 @@ import java.util.Set;
 public class OdontologoService implements IOdontologoService{
     @Autowired
      OdontologoRepository odontologoRepository;
+
     @Autowired
     ObjectMapper mapper;
 
@@ -41,12 +47,18 @@ public class OdontologoService implements IOdontologoService{
     }
 
     @Override
-    public Odontologo modificarOdontologo(OdontologoDTO odontologoDTO) {
+    public Odontologo modificarOdontologo(OdontologoDTO odontologoDTO) throws ResourceNotFoundException {
+        if(leerOdontologo(odontologoDTO.getId()) == null)
+            throw  new ResourceNotFoundException("No se puede actualizar el odontólogo ID: " + odontologoDTO.getId());
         return guardarOdontologo(odontologoDTO);
     }
 
     @Override
-    public void eliminarOdontologo(Integer id) {
+    public void eliminarOdontologo(Integer id) throws BadRequestException, ResourceNotFoundException{
+        if (id < 1)
+            throw new BadRequestException("El id del odontólogo no puede negativo");
+        if (!odontologoRepository.existsById(id))
+            throw new ResourceNotFoundException("No existe ningún odontólogo con id: " + id);
         odontologoRepository.deleteById(id);
     }
 
@@ -61,6 +73,10 @@ public class OdontologoService implements IOdontologoService{
         return odontologosDTO;
     }
 
+    @ExceptionHandler({ResourceNotFoundException.class})
+    public ResponseEntity<String>procesarErrorNotFound(ResourceNotFoundException ex){
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
+    }
 /*
 
     public Odontologo registrarOdontologo(Odontologo odontologo) {
